Complete pages
Fixed the issue I was having w/ the javascript. Pages are all done.
There's a minor spacing issue in DiveClasses.html with the headers
(specifically h3 and h4) If someone wants to go through the code and
fix it that would be great but again it's a very minor issue.